+++ doc/changelog.md
@@ -3,6 +3,7 @@
 * **0.13.3**
   - Allow dangling images to be cleaned up after build (#20)
   - Adapt order of WORKDIR and RUN when building images (#222)
+  - Allow 'build' and/or 'run' configuration to be skipped (#207)
   
 * **0.13.2**
   - "run" directives can be added to the Dockerfile (#191)